Filetypes

+++ library/programstate.py
@@ -1,4 +1,7 @@
 from typing import Union, TextIO
+from library.fasta import Fastafile
+from library.genbank import GenbankFile
+from library.record import Record
 from library.seq import PDISTANCE, JUKES_CANTOR, KIMURA_2P, PDISTANCE_GAPS, NDISTANCES, seq_distances_ufunc, seq_distances_aligned_ufunc
 import tkinter as tk
 import pandas as pd
@@ -23,8 +26,49 @@ class TabFormat(FileFormat):
     """
 
     def load_table(self, filepath_or_buffer: Union[str, TextIO]) -> pd.DataFrame:
-        return pd.read_csv(filepath_or_buffer, sep='\t').rename(
-            columns=str.casefold).rename(columns={'organism': 'species'})[['seqid', 'specimen_voucher', 'species', 'sequence']]
+        try:
+            return pd.read_csv(filepath_or_buffer, sep='\t').rename(
+                columns=str.casefold).rename(columns={'organism': 'species'})[['seqid', 'specimen_voucher', 'species', 'sequence']]
+        except KeyError as ex:
+            raise ValueError(
+                "'seqid', 'specimen_voucher', 'species' or 'organism', or 'sequence' column is missing") from ex
+
+
+class FastaFormat(FileFormat):
+    """
+    Format for fasta files
+    """
+
+    def load_table(self, filepath_or_buffer: Union[str, TextIO]) -> pd.DataFrame:
+        if isinstance(filepath_or_buffer, str):
+            with open(filepath_or_buffer) as infile:
+                return self._load_table(infile)
+        else:
+            return self._load_table(infile)
+
+    def _load_table(self, file: TextIO) -> pd.DataFrame:
+        _, records = Fastafile.read(file)
+        return pd.DataFrame(([record['seqid'], record['sequence']] for record in records()), columns=['seqid', 'sequence'])
+
+
+class GenbankFormat(FileFormat):
+    """
+    Format for fasta files
+    """
+
+    def load_table(self, filepath_or_buffer: Union[str, TextIO]) -> pd.DataFrame:
+        if isinstance(filepath_or_buffer, str):
+            with open(filepath_or_buffer) as infile:
+                return self._load_table(infile)
+        else:
+            return self._load_table(infile)
+
+    def _load_table(self, file: TextIO) -> pd.DataFrame:
+        _, records = GenbankFile.read(file)
+        try:
+            return pd.DataFrame((record._fields for record in records())).rename(columns=str.casefold).rename(columns={'organism': 'species'})[['seqid', 'specimen_voucher', 'species', 'sequence']]
+        except KeyError as ex:
+            raise ValueError(f"{str(ex)} is missing") from ex
 
 
 class ProgramState():
@@ -33,7 +77,9 @@ class ProgramState():
     """
 
     formats = dict(
-        Tabfile=TabFormat
+        Tabfile=TabFormat,
+        Fasta=FastaFormat,
+        Genbank=GenbankFormat
     )
 
     def __init__(self, root: tk.Tk) -> None:
@@ -48,11 +94,18 @@ class ProgramState():
         return ProgramState.formats[self.input_format_name.get()]()
 
     def process(self, input_file: str, output_file: str) -> None:
+        if self.input_format_name.get() == "Genbank" and self.already_aligned.get():
+            raise ValueError(
+                "'Already aligned' option is not allowed for the Genbank format.")
         with open(output_file, "w") as outfile:
-            table = self.input_format.load_table(input_file).set_index(
-                ["seqid", "specimen_voucher", "species"]).squeeze()
+            table = self.input_format.load_table(input_file)
+            sequences = table.set_index(
+                [column for column in table.columns if column != 'sequence']).squeeze()
+            if not isinstance(sequences, pd.Series):
+                raise ValueError(
+                    "Extracting sequence from the table failed for some reason")
             distance_table = make_distance_table(
-                table, self.already_aligned.get())
+                sequences, self.already_aligned.get())
             for kind in (kind for kind in range(NDISTANCES) if self.distance_options[kind].get()):
                 print(
                     f"{distances_names[kind]} between sequences", file=outfile)
